# DBMS_project/backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "DBMS Project API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "nccu-team11-dbms.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "Index",
      "description": "Index router"
    },
    {
      "name": "User",
      "description": "User router"
    },
    {
      "name": "Material",
      "description": "Material router"
    },
    {
      "name": "Product",
      "description": "Product router"
    }
  ],
  "schemes": [
    "https"
  ],
  "securityDefinitions": {},
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Index"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phoneNumber": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "註冊成功"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "使用者名稱已存在"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "登入成功"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "使用者名稱或密碼錯誤"
          }
        }
      }
    },
    "/users/getInformation": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "使用者名稱不存在"
          }
        }
      }
    },
    "/material/": {
      "get": {
        "tags": [
          "Material"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/material/getAllMaterials": {
      "post": {
        "tags": [
          "Material"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/product/": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/product/getAllProducts": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    }
  }
}